added sanity check: ignoring RT during fitting
I changed the distance function such that RT can be ignored. This makes it easy
to run EP-ABC only considering choices. The results show that drift and boundary
are correlated then (in the posterior) which makes sense, because the model
predicts the same accuracy for different combinations of drift and boundary.