initial implementation of PW-ABC
I started to implement and check piecewise ABC as described in:

White, S. R.; Kypraios, T. & Preston, S. P.
Piecewise Approximate Bayesian Computation: fast inference for discretely
observed Markov models using a factorised posterior distribution
Statistics and Computing, 2013, 25, 289-301

It seems to work for the Gaussian example and is also reasonably fast. The log
marginal likelihood is a bit off, but I suspect that this is because the
implementation ignores the first data point in the computation of the posterior.