explaining and preventing narrow posteriors
I finally had the idea to quantify the possibility of narrow posteriors when
sampling recursively. In the process I came across the issue of underestimating
standard deviations when computing unbiased estimates of variance. I have now
included a correction factor for the standard deviation (ddof=1.5) which should
reduce shrinkage of posterior standard deviations.

I also added a notebook explaining and demonstrating the issue.