changing default of doQMC to false after new insights
I further investigated the behaviour of EP-ABC in cases in which a parameter has
no meaningful effect on model predictions. This can happen in models in which a
parameter modifies a mechanism that is only in use for some values of another
parameter. If then the data indicates that the mechanism is not in use, the
modifying parameter should be non-identifiable, i.e., the posterior should equal
to prior, because there is no information in the data about this parameter.
Similarly, you might have just built a model with a useless parameter. Then
EP-ABC should indicate that by not reducing the uncertainty of the prior.

So, testing how EP-ABC behaves in such a case, I found that it consistently
shrank the prior uncertainty in the posterior even indepdently of the choice of
minacc and even consistently moved the posterior mean to smaller values. This
was for a probability parameter in [0,1] for which the posteriors strongly
preferred values <0.5 although the parameter had no effect on model predictions
and the prior was uniform.

I found that this result was due to the use of quasi-Monte-Carlo. Switching it
off lead to wider priors which were centred closer to 0.5. As expected, the
variance across EP-ABC runs increased, though, i.e., in some runs the posterior
was narrower than the prior and in others it was wider. Also as expected, I
could reduce this variance and the difference between prior and posterior by
increasing minacc from 2000 to 5000.

Given these insights, I now switched off QMC by default and expanded the
explanation of the parameter.