provide structure for parameters and their transforms
The idea is to couple parameter names with transforms so that you can
automatically plot transformed parameter values and can easily access
attributes of the transformed parameter distribution, if they are known.

Also, for the standard way of defining independent priors across parameters the
add method keeps parameter names and prior values together which should help to
prevent mistakes during the definition of priors.

The parameter container also dynamically generates a transform function that can
be accessed to transform parameter samples held in a numpy array. The idea is
that this function is as efficient as doing it by hand.