cleanup + import to package + transformed_range as instance variable
The parameters module is now loaded together with pyEPABC. Had to put tests
under the condition of __main__.

Decided that transformed_range doesn't need to be a function as I think of the
transform as unmutable after creation. May change in the future as users might
change parameters of the transform after creation which can lead to errors.