new function for estimating predictive log-likelihoods
Previously I did this using functions from the models, but when revisiting the
issues I decided that it makes most sense to have it directly in the EP-ABC
implementation.

estimate_predlik closely follows the procedures used in run_EPABC and also
shares the most important arguments. Specifically, it will keep on sampling from
the given parameter distribution for each requested trial until that trial has
at least 1 accepted sample and hence the estimated log-likelihood is greater
than infinity (unless samplemax is reached).

There does not seem a crucial difference in timing between this new
implementation and my previous procedures. Testing this for subject 4 and the
basic model of the BeeMEG experiment showed that both implementations take about
35s to estimate the posterior predictive log-likelihoods. My first run of
estimate_predlik actually took 50s, but the following two took only 35s. I don't
know where this difference came from. One possibility is that in the first run
estimate_predlik sampled more to avoid -inf, but I don't see that this can
explain the 15s difference, especially assuming that the extended sampling
happened only for one or two trials. Anyway, estimate_predlik uses much less
memory than the previous procedures, because it doesn't store any large arrays
containing simulated responses for all trials. Together with the comparable
run times and the built-in avoidance of -inf, this is a good argument for
switching the implementation.