correction: veps had a factor of 2 too much (reduced marginal likelihoods)
I mistakenly counted the RT space corresponding to the opposite choice of the
data point into the correction factor for the marginal likelihood. The
correction factor is supposed to be the normalising constant of the uniform
distribution over the data space in which you _accept_ a generated sample. In
my example I only accept responses that have the same choice as the data point
and fall within epsilon of the data RT. The correction factor should therefore
only be 2*epsilon for the space around the RT.

This should not have influenced inferences over models, because, e.g., in the
Bayes factor the constant correction factor of the evidences cancels.