bugfix + zero-transform + drop parameters
bugfix: correct number of axes in compare_pdfs

zero-transform
--------------
An alternative to the exponential transform for positive parameters. I
introduced this, because the exponential transform has issues with scale
parameters that can be expected to go towards 0.

The zero-transform maps negative values of a Gaussian random variable to 0. I
currently don't know how to represent the pdf of the transformed distribution,
because 0 becomes a value that holds actual prbobability mass. As a compromise
I don't return the density value of 0 of the underlying Gaussian, but the
probability of Gaussian values being smaller or equal to 0, i.e., the
probability mass at 0 for the transformed distribution.

For a scale that I expect to be below 1 and which may be 0 I found that the
zero-transform with mu=0, std=1 works quite well in identifying the true
values. I tested this using test_cond_inference of Pouyan's rdmeye project. For
true scale = 0 almost all posterior probability mass (e.g., 0.965) was
collected at 0 while for true scale = 0.2 the posterior had almost no mass at 0
with, e.g., mu=0.186, std=0.033. I have the impression that larger scale values
tend to be underestimated, though, which is consistent with the influence of
the prior.

drop parameters
---------------
new function drops parameters from the parameter container. For simple fitting
of reduced models after a full model has been fitted.