making it more convenient to add multiples of a parameter
When adding a parameter multiples>1 can now be requested. This will add the
requested number of parameters with the same name to the container except that
each multiple of the parameter is appended with a '_i' where i is an index
starting at 0. The added multiples of the parameter share the same transform
and mu and cov. But once added, mu and cov act independent across the added
multiples.

I introduced this to shorten the specification of parameters that only
differ, e.g., across conditions of an experiment, but should share the same
prior.